feat(wm): add basic stacking support
This commit adds support for the StackWindow, UnstackWindow and
CycleStack SocketCommands. The code is largely just migrated from
komorebi.

It looks like there isn't a public API which can be used to hide
individual windows, either via AXUIElement or something constructed
using the window id.

For now I'm just moving the stacked window off the screen and this works
well enough. I also added some cleanup which ensures that when the
process exits, windows which are off-screen will be restored to their
last known position.