feat(wm): add display reconfiguration listener
This commit adds a display reconfiguration listener which emits
notifications to a monitor reconciliator module which operates on a
background thread.

So far I'm only handling the resize notification because I have to
change the scaling when I'm recording devlogs and I don't want to have
to restart komorebi every time I do that.