feat(wm): handle monitor resizes on the main thread
Got some weird behaviour when calculating the work area offset when
resuming from sleep so I bit the bullet and added an AtomicBool flag
that gets set when we need to call into NSScreen to get the correct work
area and size after a display resizes.

The flag gets checked in the same loop on the main thread which is
checking for ctrl-c signals.