feat(wm): add reaper and input event listener
This commit adds a reaper module which gets notified whenever an
AXUIElement is invalid (so far this is notified when a set_position call
on a Window fails), and an input event listener which somewhat coarsely
emits an event whenever a mouse-up or key-up is registered.

This is to handle Apple applications like Notes and Mail which don't
fire any events at all when a window is closed.

Pretty gross. Will see if this can be improved in the future.