feat(wm): enrich monitors with serial id
This commit does some wrangling of the ioreg output to enrich monitors
with product names and serial number IDs, the latter of which is very
important to maintain compatibility with komorebi for Windows.

I can't get the equivalent of device and device_id on macOS, but these
are legacy things in komorebi for Windows so I think I'll leave them
out.

To verify monitor information, SocketMessage::MonitorInformation has
been implemented.