fix(wm): handle native tabs (mostly)
This commit adds a global list of user-define-able applications which
use native tabs (I'll expose a config option for later). This is used to
improve handling of tabs.

I have tested with Ghostty and Finder. It's possible to add tabs with
cmd+t, close them with cmd+w, and also to drag them into their own
windows.

It's also possible to cmd+n to create a new window instance in Finder,
but when doing this with Ghostty, it doesn't emit an event (I think this
is a bug in Ghostty, since it works fine with Finder).