fix(wm): handle destruction of 1st native tab
This commit implements some fixes (still not ideal) to handle the
closing of the 1st tab of an application which uses native tabs.

When the 1st tab is destroyed, the window AXUIElement gets rendered
invalid, so we have to reap that window, and then emit our own Show
event for the tab that is becoming the new 1st tab and set that as the
window.

There is an edge case now where if this happens while the window is
floating, our emitted show event will result in that window being
integrated into the layout rather than staying tiling, but I think this
is better than having a completely janky workspace and possibly having
to restart komorebi.