feat(wm): handle show events for apps that don't emit them
This commit introduces a little hack to emit a WindowManagerEvent::Show
every time the input event listener callback is triggered.

This is needed because some apps like Ghostty and Safari do not trigger
any events when a tab is split out into an independent window.

To keep things a little cleaner, a new SystemNotification::Manual
variant has been added which takes a ManualNotification variant.

The idea is to be able to use one of these whenever we manually emit a
notification to the window manager event listener to hopefully save my
sanity when we have to debug something in the future.