fix(wm): change ws on cmd+tab across ws
This is gonna drive me insane. Previous fix for something broke this.
Now we only try to focus in FocusChange on a base case if the workspace
we are on contains the target window.

This lets us fall through to the end of the handler which will try to
switch workspace if we know it's on another workspace.

Also fixed another bug around focusing the correct window on the
workspace switched to (since all the apps fire AX events when the
workspace is switched) by introducing a new ManualNotification which
lets us target the specific window that was cmd+tabbed to.