fix(wm): avoid display notification spam
Looks like macOS spams every single possible CGDisplayChangeSummaryFlag
whenever a display is added or removed, which ends up with a lot of
calls to the monitor reconciliator and a lot of locks on the wm state.

To better handle the spam this commit ensures we only send monitor
reconciliation notifications on the DesktopShapeChanged, Add and Remove
flags.