feat(wm): impl first pass of border manager
This commit introduces a very introductory attempt at a border manager.
All of the colors work, from some brief usage it seems to behave as
expected based on windows, with some caveats.

There are issues with the window layering and popups which are annoyng -
the borders always stay drawn on top of RustRover popups, for example.

Also, the window tracking is incredibly slow and laggy when you try to
move or resize using the mouse.

Probably because the implementation uses NSWindow under the hood which
according to Google is notoriously unperformant. None of the config
options have been brought over yet.

The border manager is defaulting to off for now in the absence of config
- anyone at this commit who really wants to try it out should set
BORDER_ENABLED in border_manager/mod.rs to "true".