fix(wm): improve video fullscreen handling
This commit improves browser video fullscreen handling with borders. We
are now emitting an event for
AppKitWorkspaceNotification::NSWorkspaceActiveSpaceDidChangeNotification,
even though it doesn't have a pid, by passing a dummy 0 pid.

This is really stupid and I shouldn't do this.

This lets us act with more granularity in process_event to destroy or
redraw borders when entering or exiting browser video fullscreen.

In addition, the reaper has been updated to not reap fullscreen windows
so that they can be reintegrated into the layout without issue when
fullscreen is exited.