feat(config): add window hiding position behaviour
This commit adds a new monitor configuration option
"window_hiding_position", which takes an enum with the variants
BottomLeft and BottomRight.

Until now the behaviour had been hard-coded to hide windows at the
bottom left of every monitor.

If no value is given on a monitor configuration object, this option will
default to BottomLeft.