fix(wm): add hack to manage activity monitor on start
I was running into this a lot when debugging the WindowServer memory
ballooning issue fixed in the last commit - the title returned by
CGWindowListCopyWindowInfo and the title returned by kAXWindowsAttribute
for Activity Monitor differs, so when restarting komorebi, if Activity
Monitor is open, it wouldn't get managed.

Added a little hack as a workaround for now.