fix(wm): account for partially-shown scrolling windows
This commit makes some small tweaks to account for the hiding and
showing of partially visible windows on the Scrolling layout.

This can happen when resize dimensions have been applied which leaves
only a certain percentage of the window within the work area.

If a window is only partially within the work area, the call to hide
will silently fail.

In order to deal with this there is some jank added when hiding
workspaces and updating workspaces to see if we are leaving a Scrolling
workspace with a partially-visible window, and if we are, to reposition
that window to be fully on within the work area before hiding it.