fix(wm): expand hack from 4d25d4e for chrome
This commit expands the hack introduced in 4d25d4e which deals with
inconsistent window titles returned by apps from different macOS APIs.
The same issue that was happening with Activity Monitor was also
happening with Google Chrome, so this hack has been updated to a more
general "contains" check.