fix(wm): explicitly init and process nsapplication events
For reasons unknown, CFRunLoop::run_in_mode no longer processes
NSApplication events on Tahoe.

Display callbacks seem to now be more tightly coupled to the
NSApplication event loop, so we initialize NSApplication with
sharedApplication as early as possible to establish a window server
connection, and then in our main loop we poll app.nextEvent every 0.1s
for NSApplication events, which is where the
DisplayReconfigurationListener callback will finally kick in again.