fix(wm): consume more appkit events and make border manager reconcile focus
This commit ensures that NSWorkspaceDidActivateApplicationNotification
and NSWorkspaceDidDeactivateApplicationNotification get emitted not just
for the loginwindow but for all apps with a valid name.

This helps us to catch events from windows which don't emit
AXWindowCreated consistently like Mail, Calendar, etc.

Additionally, if a window has been closed with cmd+w but the application
is still active, the border manager will pick a window container to mark
as the new focused window container, and call window.focus() on the
selected window to align the focus with the focused border color.